Remove all non-essential functions from _ast_util.py
As this is an old module that keeps changing with new Python
releases, combined with the fact that we no longer use "import *",
remove all unused functions from this module so that the list
of imports we are maintaining is much shorter.

Python 3.8 is also added to tox.ini with more adjustments to
test_exceptions to accommodate for quoting differences.

Fixes: #296
Change-Id: I0cea7d21ecb2be9335e0845ae7fbd06d468a137b