Correctly track line numbers for multi-line code blocks
Up till now, exceptions in <% … %> or <%!  … %> code blocks would have their
errors reported in the opening line when formatting a stack trace.  With
these changes, every line of a code block gets an entry in the line map,
which allows attributing them correctly.

Exceptions in <%! … %> blocks which get raised while loading the module are
still not reported correctly.  At that time, the metadata infrastructure for
exception reporting does not exist yet, so the stack trace will reference to
the raw generated code and not get rewritten.