create a mako main shebang entry point
That way, mako can directly evaluate template files,
or can be used as a file interpreter.

When using `#!/usr/bin/env -S mako-render -s --` or
`#!/usr/bin/env -S python3 -m mako -s --` as file shebang,
a file can be templated directly, and behaves similar to a PHP script.
With this commit, Python can basically be run as PyHP :)

This is very useful for example for generating static config files,
especially in combination with some `execfs` which executes files
when they are read. That way one can even programatically customize
the a precious `xorg.conf`.